+++ packages/menu/Menu.vue
@@ -1,85 +1,46 @@
-<script type="text/jsx">
-import Menubar from 'element-ui/src/utils/menu/aria-menubar';
-import { addClass, removeClass, hasClass } from 'element-ui/src/utils/dom';
-import { computed, getCurrentInstance, onMounted, provide, ref, toRefs, watch, h , Transition } from 'vue';
-import {useEmitter} from 'element-ui/src/use/emitter';
-import {useItems} from "./src/menu-use";
+<script>
+import Menubar from 'element-ui/src/utils/menu/aria-menubar'
+import {
+  computed,
+  getCurrentInstance,
+  onMounted,
+  provide,
+  ref,
+  toRefs,
+  watch
+} from 'vue'
+import { useEmitter } from 'element-ui/src/use/emitter'
+import { useItems } from './src/menu-use'
+import MenuCollapseTransition from './MenuCollapseTransition.vue'
 
 export default {
   name: 'ElMenu',
   componentName: 'ElMenu',
-  render () {
+  render() {
     const component = (
       <ul
         role="menubar"
-        key={ +this.collapse }
+        key={+this.collapse}
         style={{ backgroundColor: this.backgroundColor || '' }}
         class={{
           'el-menu--horizontal': this.mode === 'horizontal',
           'el-menu--collapse': this.collapse,
-          "el-menu": true
+          'el-menu': true
         }}
       >
-        { this.$slots.default() }
+        {this.$slots.default()}
       </ul>
-    );
+    )
     if (this.collapseTransition) {
       return (
-        <el-menu-collapse-transition>
-          { component }
-        </el-menu-collapse-transition>
-      );
+        <el-menu-collapse-transition>{component}</el-menu-collapse-transition>
+      )
     } else {
-      return component;
+      return component
     }
-
   },
   components: {
-    'el-menu-collapse-transition': {
-      render(ctx) {
-        const data = {
-          mode: 'out-in',
-          onBeforeEnter(el) {
-            el.style.opacity = 0.2;
-          },
-
-          onEnter(el) {
-            addClass(el, 'el-opacity-transition');
-            el.style.opacity = 1;
-          },
-
-          onAfterEnter(el) {
-            removeClass(el, 'el-opacity-transition');
-            el.style.opacity = '';
-          },
-
-          onBeforeLeave(el) {
-            if (!el.dataset) el.dataset = {};
-
-            if (hasClass(el, 'el-menu--collapse')) {
-              removeClass(el, 'el-menu--collapse');
-              el.dataset.oldOverflow = el.style.overflow;
-              el.dataset.scrollWidth = el.clientWidth;
-              addClass(el, 'el-menu--collapse');
-            } else {
-              addClass(el, 'el-menu--collapse');
-              el.dataset.oldOverflow = el.style.overflow;
-              el.dataset.scrollWidth = el.clientWidth;
-              removeClass(el, 'el-menu--collapse');
-            }
-
-            el.style.width = el.scrollWidth + 'px';
-            el.style.overflow = 'hidden';
-          },
-
-          onLeave(el) {
-            addClass(el, 'horizontal-collapse-transition');
-            el.style.width = el.dataset.scrollWidth + 'px';
-          }
-        };
-        return h(Transition, data, ctx.$slots);
-      }
-    }
+    'el-menu-collapse-transition': MenuCollapseTransition
   },
 
   props: {
@@ -101,7 +62,7 @@ export default {
       type: String,
       default: 'hover'
     },
-    collapse:  {
+    collapse: {
       type: Boolean,
       default: false
     },
@@ -116,8 +77,8 @@ export default {
       default: true
     }
   },
-  emits:["close","open","select","on"],
-  setup(props,{ emit }){
+  emits: ['close', 'open', 'select', 'on'],
+  setup(props, { emit }) {
     const {
       defaultActive,
       defaultOpeneds,
@@ -126,175 +87,187 @@ export default {
       backgroundColor,
       mode,
       router
-    } = toRefs(props);
-    const instance = getCurrentInstance();
-    const {broadcast, on} = useEmitter()
-    const activeIndex = ref(defaultActive.value);
-    const openedMenus = ref((defaultOpeneds.value && !collapse.value) ? defaultOpeneds.value.slice(0) : []);
+    } = toRefs(props)
+    const instance = getCurrentInstance()
+    const { broadcast, on } = useEmitter()
+    const activeIndex = ref(defaultActive.value)
+    const openedMenus = ref(
+      defaultOpeneds.value && !collapse.value
+        ? defaultOpeneds.value.slice(0)
+        : []
+    )
     const {
-        items,
-        itemsInstance,
-        submenusInstance,
-        addItem,
-        removeItem,
-        addSubmenu,
-        removeSubmenu
-      } = useItems();
-    const updateActiveIndex=(val)=>{
-      let item = items[val] || items[activeIndex.value] || items[defaultActive.value];
-      item = itemsInstance[item];
+      items,
+      itemsInstance,
+      submenusInstance,
+      addItem,
+      removeItem,
+      addSubmenu,
+      removeSubmenu
+    } = useItems()
+    const updateActiveIndex = (val) => {
+      let item =
+        items[val] || items[activeIndex.value] || items[defaultActive.value]
+      item = itemsInstance[item]
       if (item) {
-        activeIndex.value = item.ctx.index;
-        initOpenedMenu();
+        activeIndex.value = item.ctx.index
+        initOpenedMenu()
       } else {
-        activeIndex.value = null;
+        activeIndex.value = null
       }
     }
-    const getColorChannels = (color)=>{
-      color = color.replace('#', '');
+    const getColorChannels = (color) => {
+      color = color.replace('#', '')
       if (/^[0-9a-fA-F]{3}$/.test(color)) {
-        color = color.split('');
+        color = color.split('')
         for (let i = 2; i >= 0; i--) {
-          color.splice(i, 0, color[i]);
+          color.splice(i, 0, color[i])
         }
-        color = color.join('');
+        color = color.join('')
       }
       if (/^[0-9a-fA-F]{6}$/.test(color)) {
         return {
           red: parseInt(color.slice(0, 2), 16),
           green: parseInt(color.slice(2, 4), 16),
           blue: parseInt(color.slice(4, 6), 16)
-        };
+        }
       } else {
         return {
           red: 255,
           green: 255,
           blue: 255
-        };
+        }
       }
-    };
-    const mixColor = (color, percent)=>{
-      let { red, green, blue } = getColorChannels(color);
-      if (percent > 0) { // shade given color
-        red *= 1 - percent;
-        green *= 1 - percent;
-        blue *= 1 - percent;
-      } else { // tint given color
-        red += (255 - red) * percent;
-        green += (255 - green) * percent;
-        blue += (255 - blue) * percent;
-      }
-      return `rgb(${ Math.round(red) }, ${ Math.round(green) }, ${ Math.round(blue) })`;
     }
-    const openMenu = (index, indexPath)=>{
-      if (openedMenus.value.indexOf(index) !== -1) return;
+    const mixColor = (color, percent) => {
+      let { red, green, blue } = getColorChannels(color)
+      if (percent > 0) {
+        // shade given color
+        red *= 1 - percent
+        green *= 1 - percent
+        blue *= 1 - percent
+      } else {
+        // tint given color
+        red += (255 - red) * percent
+        green += (255 - green) * percent
+        blue += (255 - blue) * percent
+      }
+      return `rgb(${Math.round(red)}, ${Math.round(green)}, ${Math.round(
+        blue
+      )})`
+    }
+    const openMenu = (index, indexPath) => {
+      if (openedMenus.value.indexOf(index) !== -1) return
       // 将不在该菜单路径下的其余菜单收起
       // collapse all menu that are not under current menu item
       if (uniqueOpened.value) {
-        openedMenus.value = openedMenus.value.filter(index => {
-          return indexPath.indexOf(index) !== -1;
-        });
+        openedMenus.value = openedMenus.value.filter((index) => {
+          return indexPath.indexOf(index) !== -1
+        })
       }
-      openedMenus.value.push(index);
+      openedMenus.value.push(index)
     }
-    const closeMenu = (index)=>{
-      const i = openedMenus.value.indexOf(index);
+    const closeMenu = (index) => {
+      const i = openedMenus.value.indexOf(index)
       if (i !== -1) {
-        openedMenus.value.splice(i, 1);
+        openedMenus.value.splice(i, 1)
       }
     }
-    const handleSubmenuClick = (submenu)=>{
-      const { index, indexPath } = submenu.ctx;
-      const isOpened = openedMenus.value.indexOf(index) !== -1;
+    const handleSubmenuClick = (submenu) => {
+      const { index, indexPath } = submenu.ctx
+      const isOpened = openedMenus.value.indexOf(index) !== -1
       if (isOpened) {
-        closeMenu(index);
-        emit('close', index, indexPath);
+        closeMenu(index)
+        emit('close', index, indexPath)
       } else {
-        openMenu(index, indexPath);
-        emit('open', index, indexPath);
+        openMenu(index, indexPath)
+        emit('open', index, indexPath)
       }
     }
-    const handleItemClick = (item)=>{
-      const { index, indexPath } = item.ctx;
-      const oldActiveIndex = activeIndex.value;
-      const hasIndex = index !== null;
+    const handleItemClick = (item) => {
+      const { index, indexPath } = item.ctx
+      const oldActiveIndex = activeIndex.value
+      const hasIndex = index !== null
       if (hasIndex) {
-        activeIndex.value = index;
+        activeIndex.value = index
       }
-      emit('select', index, indexPath, item);
+      emit('select', index, indexPath, item)
       if (mode.value === 'horizontal' || collapse.value) {
-        openedMenus.value.length = 0;
+        openedMenus.value.length = 0
       }
       if (router.value && hasIndex) {
         routeToItem(item, (error) => {
-          activeIndex.value = oldActiveIndex;
+          activeIndex.value = oldActiveIndex
           if (error) {
             // vue-router 3.1.0+ push/replace cause NavigationDuplicated error
             // https://github.com/ElemeFE/element/issues/17044
             if (error.name === 'NavigationDuplicated') return
             console.error(error)
           }
-        });
+        })
       }
     }
-    const initOpenedMenu = ()=>{
-      const index = activeIndex.value;
-      const activeItem = itemsInstance[index];
-      if (!activeItem || mode.value === 'horizontal' || collapse.value) return;
-      const indexPath = activeItem.ctx.indexPath;
+    const initOpenedMenu = () => {
+      const index = activeIndex.value
+      const activeItem = itemsInstance[index]
+      if (!activeItem || mode.value === 'horizontal' || collapse.value) return
+      const indexPath = activeItem.ctx.indexPath
       // 展开该菜单项的路径上所有子菜单
       // expand all submenus of the menu item
-      indexPath.forEach(index => {
-        const submenu = submenusInstance[index];
-        submenu && openMenu(index, submenu.ctx.indexPath);
-      });
+      indexPath.forEach((index) => {
+        const submenu = submenusInstance[index]
+        submenu && openMenu(index, submenu.ctx.indexPath)
+      })
     }
-    const routeToItem = (item, onError)=>{
-      const route = item.route || item.index;
+    const routeToItem = (item, onError) => {
+      const route = item.route || item.index
       try {
-        instance.ctx.$router.push(route, () => {}, onError);
+        instance.ctx.$router.push(route, () => {}, onError)
       } catch (e) {
-        console.error(e);
+        console.error(e)
       }
     }
-    const open = (index)=>{
-      const { indexPath } = submenusInstance[index.toString()];
-      indexPath.forEach(i => openMenu(i, indexPath));
+    const open = (index) => {
+      const { indexPath } = submenusInstance[index.toString()]
+      indexPath.forEach((i) => openMenu(i, indexPath))
     }
-    const close = (index)=>{
-      closeMenu(index);
+    const close = (index) => {
+      closeMenu(index)
     }
-    const hoverBackground = computed(()=>{
-      return backgroundColor.value ? mixColor(backgroundColor.value, 0.2) : '';
+    const hoverBackground = computed(() => {
+      return backgroundColor.value ? mixColor(backgroundColor.value, 0.2) : ''
     })
-    const isMenuPopup = computed(()=>{
-      return mode.value === 'horizontal' || (mode.value === 'vertical' && collapse.value);
-    });
-    watch(defaultActive,(value)=>{
-      if(!items[value]){
+    const isMenuPopup = computed(() => {
+      return (
+        mode.value === 'horizontal' ||
+        (mode.value === 'vertical' && collapse.value)
+      )
+    })
+    watch(defaultActive, (value) => {
+      if (!items[value]) {
         activeIndex.value = null
       }
       updateActiveIndex(value)
-    });
-    watch(defaultOpeneds,(value)=>{
+    })
+    watch(defaultOpeneds, (value) => {
       if (!collapse.value) {
-        openedMenus.value = value;
+        openedMenus.value = value
       }
-    });
-    watch(collapse,(value)=>{
-      if (value) openedMenus.value = [];
-      broadcast('toggle-collapse', value);
-    });
-    onMounted(()=>{
-        initOpenedMenu();
-        on('item-click', handleItemClick);
-        on('submenu-click', handleSubmenuClick);
-        if (mode.value === 'horizontal') {
-            new Menubar(instance.ctx.$el); // eslint-disable-line
-        }
-        watch(items, updateActiveIndex);
-    });
-    provide("rootMenu",instance);
+    })
+    watch(collapse, (value) => {
+      if (value) openedMenus.value = []
+      broadcast('toggle-collapse', value)
+    })
+    onMounted(() => {
+      initOpenedMenu()
+      on('item-click', handleItemClick)
+      on('submenu-click', handleSubmenuClick)
+      if (mode.value === 'horizontal') {
+        new Menubar(instance.ctx.$el)
+      }
+      watch(items, updateActiveIndex)
+    })
+    provide('rootMenu', instance)
     return {
       addItem,
       removeItem,
@@ -310,5 +283,5 @@ export default {
       close
     }
   }
-};
+}
 </script>