+++ packages/page-header/PageHeader.vue
@@ -25,6 +25,7 @@ export default {
     },
     content: String
   },
+  emits: ['back'],
   setup(props, ctx) {
     const handleClick = (evt) => {
       ctx.emit('back', evt)